pyproject.toml
[project]
name = "chunking-benchmark"
version = "0.1.0"
description = "A comprehensive chunking benchmark with support for multiple file formats and vector storage"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.4",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "langchain-text-splitters>=0.0.1",
    "langchain-community>=0.1.0",
    "langchain-openai>=0.0.5",
    "tiktoken>=0.5.0",
    "tqdm>=4.66.0",
    "pandas>=2.2.0",
    "ragas>=0.1.0",
    "openai>=1.0.0",
    "datasets>=2.16.0",
    "pinecone>=7.3.0",
    "torch>=2.0.0",
    "transformers>=4.43.4",
    "peft>=0.17.1",
    "einops>=0.8.1",
]
